fix(web): move last-seen activity fetch out of shared useUserManagement hook
The /api/users/activity fetch (superadmin-only) was living in
useUserManagement, but that hook is also used by ManageSitesDialog on
non-superadmin pages (roosts/dashboard/deployments/logs) — so the fetch
fired there and 403'd, and the browser logged that as a console error,
tripping the empty-roost-state e2e console-error guard. Moved the fetch to
the superadmin-only admin/users page (mirroring the deletions fetch); the
shared hook is back to user-list + mutations + deleted-aware counts only.

Also fix the user-mgmt e2e assertion to match the lowercased "you" pill.
