+++ web/__tests__/hooks/useUserManagement.test.ts
@@ -2,8 +2,8 @@
  * @jest-environment jsdom
  *
  * Unit tests for `useUserManagement` — the realtime user list, soft-delete
- * flagging, role counts that exclude deleted users, and the non-fatal
- * `/api/users/activity` merge.
+ * flagging, and role counts that exclude deleted users. (Last-seen/activity
+ * is fetched page-side in app/admin/users, not by this shared hook.)
  */
 import { renderHook, waitFor } from '@testing-library/react';
 
@@ -48,7 +48,6 @@ beforeEach(() => {
 
 afterEach(() => {
   jest.restoreAllMocks();
-  delete (global as { fetch?: unknown }).fetch;
 });
 
 describe('useUserManagement — soft delete', () => {
@@ -57,10 +56,9 @@ describe('useUserManagement — soft delete', () => {
       activeDoc('u-active', 'member'),
       deletedDoc('u-deleted', 'admin', 1700000000000),
     ];
-    global.fetch = jest.fn().mockResolvedValue({ ok: true, json: async () => ({ activity: {} }) });
 
     const { result } = renderHook(() => useUserManagement());
-    await waitFor(() => expect(global.fetch).toHaveBeenCalled());
+    await waitFor(() => expect(result.current.users).toHaveLength(2));
 
     const active = result.current.users.find((u) => u.uid === 'u-active');
     const deleted = result.current.users.find((u) => u.uid === 'u-deleted');
@@ -79,10 +77,9 @@ describe('useUserManagement — getUserCounts', () => {
       deletedDoc('u-del-1', 'member', 1700000000000),
       deletedDoc('u-del-2', 'admin', 1700000000001),
     ];
-    global.fetch = jest.fn().mockResolvedValue({ ok: true, json: async () => ({ activity: {} }) });
 
     const { result } = renderHook(() => useUserManagement());
-    await waitFor(() => expect(global.fetch).toHaveBeenCalled());
+    await waitFor(() => expect(result.current.users).toHaveLength(5));
 
     const counts = result.current.getUserCounts();
     expect(counts.total).toBe(3);
@@ -90,57 +87,5 @@ describe('useUserManagement — getUserCounts', () => {
     expect(counts.admins).toBe(1);
     expect(counts.members).toBe(1);
     expect(counts.deleted).toBe(2);
-  });
-});
-
-describe('useUserManagement — activity', () => {
-  it('populates activity from /api/users/activity merged by uid', async () => {
-    snapshotDocs = [activeDoc('u-1', 'member'), activeDoc('u-2', 'admin')];
-    const activityPayload = {
-      'u-1': { lastSignInTime: '2026-05-01T00:00:00Z', lastRefreshTime: null, disabled: false },
-      'u-2': { lastSignInTime: null, lastRefreshTime: '2026-05-02T00:00:00Z', disabled: true },
-    };
-    const fetchMock = jest
-      .fn()
-      .mockResolvedValue({ ok: true, json: async () => ({ activity: activityPayload }) });
-    global.fetch = fetchMock;
-
-    const { result } = renderHook(() => useUserManagement());
-
-    await waitFor(() => {
-      expect(result.current.activity['u-1']).toBeDefined();
-    });
-    expect(fetchMock).toHaveBeenCalledWith('/api/users/activity');
-    expect(result.current.activity['u-1'].lastSignInTime).toBe('2026-05-01T00:00:00Z');
-    expect(result.current.activity['u-2'].disabled).toBe(true);
-  });
-
-  it('leaves users intact and activity empty when the fetch fails (non-fatal)', async () => {
-    snapshotDocs = [activeDoc('u-1', 'member')];
-    const consoleSpy = jest.spyOn(console, 'error').mockImplementation(() => {});
-    global.fetch = jest.fn().mockRejectedValue(new Error('network down'));
-
-    const { result } = renderHook(() => useUserManagement());
-
-    await waitFor(() => {
-      expect(consoleSpy).toHaveBeenCalled();
-    });
-    expect(result.current.users).toHaveLength(1);
-    expect(result.current.users[0].uid).toBe('u-1');
-    expect(result.current.activity).toEqual({});
-  });
-
-  it('treats a non-ok response as non-fatal and keeps activity empty', async () => {
-    snapshotDocs = [activeDoc('u-1', 'member')];
-    const consoleSpy = jest.spyOn(console, 'error').mockImplementation(() => {});
-    global.fetch = jest.fn().mockResolvedValue({ ok: false, status: 500, json: async () => ({}) });
-
-    const { result } = renderHook(() => useUserManagement());
-
-    await waitFor(() => {
-      expect(consoleSpy).toHaveBeenCalled();
-    });
-    expect(result.current.users).toHaveLength(1);
-    expect(result.current.activity).toEqual({});
   });
 });
